refactor(audio): drop dead level payload and finish the tap migration
Three cleanups from a /simplify pass over the sidecar audio-level work,
all the same theme — complete the move off the renderer-side AnalyserNode
tap and stop carrying data nothing reads:

- Remove peak/rms from the audio_level event end-to-end (MixedAudioFrame →
  Event::AudioLevel → AudioLevelEvent). No consumer ever read them; the
  meter uses only bands.
- Compute the level FFT lazily. The mix path runs every 20 ms frame but
  emission is throttled to ~20 Hz, so the FFT was computed and discarded on
  most frames. analyze_levels() now runs only when an event is actually due.
- Delete the orphaned AudioVisualizerTap (unwired once levels moved into the
  sidecar) and its test, plus the dead visualizer plumbing in
  AudioCaptureStream. Move the still-shared BAND_COUNT + AudioBandReader into
  a neutral audio-bands module that the meter and ribbon import.
